+++ packages/tooling/fast-tooling/src/web-components/html-render/html-render.styles.ts
@@ -4,8 +4,11 @@ export const htmlRenderStyles = (context, definition) => css`
     .container {
         width: 100%;
         height: 100%;
-        padding: 40px;
+        padding: 0;
         box-sizing: border-box;
+    }
+    .container.interactive {
+        padding: 40px;
     }
     .html-render {
         width: 100%;