+++ packages/tooling/fast-tooling-react/CHANGELOG.md
@@ -1,8 +1,54 @@
 # Change Log - @microsoft/fast-tooling-react
 
-This log was last generated on Tue, 29 Jun 2021 07:14:52 GMT and should not be manually modified.
+This log was last generated on Tue, 13 Jul 2021 07:14:52 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.9.0
+
+Tue, 13 Jul 2021 07:14:52 GMT
+
+### Minor changes
+
+- update the monaco adapter, html render and html render layer and react Form to update their stored schema dictionaries when the schema dictionary has new schemas added to it (7559015+janechu@users.noreply.github.com)
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.2.1 (ben@platform5.ch)
+- Bump @microsoft/fast-tooling to v0.24.0 (7559015+janechu@users.noreply.github.com)
+- Bump @microsoft/fast-foundation to v2.4.0 (roeisenb@microsoft.com)
+
+## 2.8.26
+
+Sun, 11 Jul 2021 07:15:10 GMT
+
+### Patches
+
+- Adding support for preview mode styles to the viewer (44823142+williamw2@users.noreply.github.com)
+
+## 2.8.25
+
+Wed, 07 Jul 2021 07:19:51 GMT
+
+### Patches
+
+- Bump @microsoft/fast-tooling to v0.23.6 (scomea@microsoft.com)
+
+## 2.8.24
+
+Sun, 04 Jul 2021 07:15:28 GMT
+
+### Patches
+
+- Bump @microsoft/fast-tooling to v0.23.5 (john.kreitlow@microsoft.com)
+
+## 2.8.23
+
+Thu, 01 Jul 2021 07:15:57 GMT
+
+### Patches
+
+- Bump @microsoft/fast-tooling to v0.23.4 (nicholasrice@users.noreply.github.com)
 
 ## 2.8.22